src/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai-whisper
ffmpeg-python
tqdm